requirements.txt
streamlit
python-dotenv
pytesseract
Pillow